Default NameIdFormat to unspecified
Name id format is pretty fundamental to SAML2 workflow but rather than
blowing up and throwing an exception we will use a default of
`urn:oasis:names:tc:SAML:1.1:nameid-format:unspecified`. The consequence
of this change is that using the format to build up things like SAML
sign out requests will may not actually provide a format that will be
accepted by the SAML identity provider.